refactor navigation components and add footer section for blog layout
