+++ components/blog/base/navigation/NavigationSimpleButton.vue
