+++ pages/blog/[...slug].vue
